+++ packages/contracts/api/index.ts
@@ -15,3 +15,4 @@ export * from '../schemas/media';
 export * from '../schemas/password-recovery';
 export * from '../schemas/settings';
 export * from '../schemas/users';
+export * from '../schemas/webhooks';